+++ crates/full-node/sov-ledger-apis/src/lib.rs
@@ -1,8 +1,8 @@
+mod metrics;
+
 use std::collections::HashMap;
-use std::io::Write;
 use std::marker::PhantomData;
 use std::ops::Range;
-use std::sync::atomic::{AtomicU64, Ordering};
 
 use axum::extract::{Request, State, WebSocketUpgrade};
 use axum::http::StatusCode;
@@ -16,7 +16,6 @@ use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
 use sov_db::schema::types::{BatchNumber, EventNumber, TxNumber};
-use sov_metrics::Metric;
 use sov_modules_api::da::Time;
 use sov_modules_api::rpc::LedgerStateProviderExt;
 pub use sov_modules_api::ApiTxEffect as TxEffect;
@@ -38,141 +37,9 @@ use sov_rollup_interface::node::ledger_api::{
 use sov_rollup_interface::stf::TxReceiptContents;
 use tokio::sync::watch;
 
+use crate::metrics::{track_finalized_ws_replay, LedgerWsConnectionGuard};
+
 type PathMap = Path<HashMap<String, NumberOrHash>>;
-
-const LEDGER_AGGREGATED_PROOFS_WS_ROUTE: &str = "/ledger/aggregated-proofs/latest/ws";
-const LEDGER_SLOT_EVENTS_WS_ROUTE: &str = "/ledger/slots/latest/events/ws";
-const LEDGER_SLOTS_LATEST_WS_ROUTE: &str = "/ledger/slots/latest/ws";
-const LEDGER_SLOTS_FINALIZED_WS_ROUTE: &str = "/ledger/slots/finalized/ws";
-const NO_QUERY_MODE_LABEL: &str = "none";
-
-static LEDGER_AGGREGATED_PROOFS_WS_CONNECTIONS: AtomicU64 = AtomicU64::new(0);
-static LEDGER_SLOT_EVENTS_WS_CONNECTIONS: AtomicU64 = AtomicU64::new(0);
-static LEDGER_SLOTS_LATEST_WS_COMPACT_CONNECTIONS: AtomicU64 = AtomicU64::new(0);
-static LEDGER_SLOTS_LATEST_WS_STANDARD_CONNECTIONS: AtomicU64 = AtomicU64::new(0);
-static LEDGER_SLOTS_LATEST_WS_FULL_CONNECTIONS: AtomicU64 = AtomicU64::new(0);
-static LEDGER_SLOTS_FINALIZED_WS_COMPACT_CONNECTIONS: AtomicU64 = AtomicU64::new(0);
-static LEDGER_SLOTS_FINALIZED_WS_STANDARD_CONNECTIONS: AtomicU64 = AtomicU64::new(0);
-static LEDGER_SLOTS_FINALIZED_WS_FULL_CONNECTIONS: AtomicU64 = AtomicU64::new(0);
-
-struct LedgerWsConnectionGuard {
-    route: &'static str,
-    query_mode: &'static str,
-    counter: &'static AtomicU64,
-}
-
-impl LedgerWsConnectionGuard {
-    fn new(route: &'static str, query_mode: &'static str, counter: &'static AtomicU64) -> Self {
-        let active_connections = counter.fetch_add(1, Ordering::Relaxed) + 1;
-        track_ledger_ws_connections(route, query_mode, active_connections);
-        Self {
-            route,
-            query_mode,
-            counter,
-        }
-    }
-}
-
-impl Drop for LedgerWsConnectionGuard {
-    fn drop(&mut self) {
-        let previous = self.counter.fetch_sub(1, Ordering::Relaxed);
-        let active_connections = previous.saturating_sub(1);
-        track_ledger_ws_connections(self.route, self.query_mode, active_connections);
-    }
-}
-
-#[derive(Debug)]
-struct FinalizedWsReplayMetrics {
-    query_mode: &'static str,
-    replayed_slots: u64,
-}
-
-impl Metric for FinalizedWsReplayMetrics {
-    fn measurement_name(&self) -> &'static str {
-        "sov_rollup_ledger_finalized_ws_replay"
-    }
-
-    fn serialize_for_telegraf(&self, buffer: &mut Vec<u8>) -> std::io::Result<()> {
-        write!(
-            buffer,
-            "{},query_mode={} replayed_slots={}",
-            self.measurement_name(),
-            self.query_mode,
-            self.replayed_slots,
-        )
-    }
-}
-
-fn query_mode_label(query_mode: QueryMode) -> &'static str {
-    match query_mode {
-        QueryMode::Compact => "compact",
-        QueryMode::Standard => "standard",
-        QueryMode::Full => "full",
-    }
-}
-
-fn track_finalized_ws_replay(query_mode: QueryMode, replayed_slots: u64) {
-    sov_metrics::track_metrics(|tracker| {
-        tracker.submit(FinalizedWsReplayMetrics {
-            query_mode: query_mode_label(query_mode),
-            replayed_slots,
-        });
-    });
-}
-
-#[derive(Debug)]
-struct LedgerWsConnectionsMetric {
-    route: &'static str,
-    query_mode: &'static str,
-    active_connections: u64,
-}
-
-impl Metric for LedgerWsConnectionsMetric {
-    fn measurement_name(&self) -> &'static str {
-        "sov_rollup_ledger_ws_connections"
-    }
-
-    fn serialize_for_telegraf(&self, buffer: &mut Vec<u8>) -> std::io::Result<()> {
-        write!(
-            buffer,
-            "{},route={},query_mode={} active_connections={}",
-            self.measurement_name(),
-            self.route,
-            self.query_mode,
-            self.active_connections,
-        )
-    }
-}
-
-fn track_ledger_ws_connections(
-    route: &'static str,
-    query_mode: &'static str,
-    active_connections: u64,
-) {
-    sov_metrics::track_metrics(|tracker| {
-        tracker.submit(LedgerWsConnectionsMetric {
-            route,
-            query_mode,
-            active_connections,
-        });
-    });
-}
-
-fn head_ws_counter(query_mode: QueryMode) -> &'static AtomicU64 {
-    match query_mode {
-        QueryMode::Compact => &LEDGER_SLOTS_LATEST_WS_COMPACT_CONNECTIONS,
-        QueryMode::Standard => &LEDGER_SLOTS_LATEST_WS_STANDARD_CONNECTIONS,
-        QueryMode::Full => &LEDGER_SLOTS_LATEST_WS_FULL_CONNECTIONS,
-    }
-}
-
-fn finalized_ws_counter(query_mode: QueryMode) -> &'static AtomicU64 {
-    match query_mode {
-        QueryMode::Compact => &LEDGER_SLOTS_FINALIZED_WS_COMPACT_CONNECTIONS,
-        QueryMode::Standard => &LEDGER_SLOTS_FINALIZED_WS_STANDARD_CONNECTIONS,
-        QueryMode::Full => &LEDGER_SLOTS_FINALIZED_WS_FULL_CONNECTIONS,
-    }
-}
 
 /// Maximum value of `page[cursor]` accepted on the prefix path of `list_events`.
 /// Bounds per-request work since the current implementation re-reads everything
@@ -842,11 +709,7 @@ where
         ws: WebSocketUpgrade,
     ) -> impl IntoResponse {
         ws.on_upgrade(|socket| async move {
-            let _connection_guard = LedgerWsConnectionGuard::new(
-                LEDGER_SLOT_EVENTS_WS_ROUTE,
-                NO_QUERY_MODE_LABEL,
-                &LEDGER_SLOT_EVENTS_WS_CONNECTIONS,
-            );
+            let _connection_guard = LedgerWsConnectionGuard::slot_events();
             let subscription = state
                 .ledger
                 .subscribe_slots()
@@ -885,11 +748,7 @@ where
         ws: WebSocketUpgrade,
     ) -> impl IntoResponse {
         ws.on_upgrade(|socket| async move {
-            let _connection_guard = LedgerWsConnectionGuard::new(
-                LEDGER_AGGREGATED_PROOFS_WS_ROUTE,
-                NO_QUERY_MODE_LABEL,
-                &LEDGER_AGGREGATED_PROOFS_WS_CONNECTIONS,
-            );
+            let _connection_guard = LedgerWsConnectionGuard::aggregated_proofs();
             let subscription = state.ledger.subscribe_proof_saved().map(|data| {
                 AggregatedProof::try_from(data).map_err(|e| {
                     tracing::error!(
@@ -916,11 +775,7 @@ where
         let ledger = state.ledger;
 
         ws.on_upgrade(move |socket| async move {
-            let _connection_guard = LedgerWsConnectionGuard::new(
-                LEDGER_SLOTS_LATEST_WS_ROUTE,
-                query_mode_label(query_mode),
-                head_ws_counter(query_mode),
-            );
+            let _connection_guard = LedgerWsConnectionGuard::head(query_mode);
             let subscription = ledger
                 .subscribe_slots()
                 .then(|slot_num| {
@@ -958,11 +813,7 @@ where
         let ledger = state.ledger;
 
         ws.on_upgrade(move |socket| async move {
-            let _connection_guard = LedgerWsConnectionGuard::new(
-                LEDGER_SLOTS_FINALIZED_WS_ROUTE,
-                query_mode_label(query_mode),
-                finalized_ws_counter(query_mode),
-            );
+            let _connection_guard = LedgerWsConnectionGuard::finalized(query_mode);
             let initial_slot = match ledger.get_latest_finalized_slot_number().await {
                 Ok(s) => s,
                 Err(error) => {
